+++ axum/src/extract/matched_path.rs
@@ -148,7 +148,7 @@ mod tests {
                 "/:key",
                 get(|path: MatchedPath| async move { path.as_str().to_owned() }),
             )
-            .nest("/api", api.into_service())
+            .nest("/api", api)
             .nest(
                 "/public",
                 Router::new()
@@ -156,10 +156,9 @@ mod tests {
                     // have to set the middleware here since otherwise the
                     // matched path is just `/public/*` since we're nesting
                     // this router
-                    .layer(layer_fn(SetMatchedPathExtension))
-                    .into_service(),
+                    .layer(layer_fn(SetMatchedPathExtension)),
             )
-            .nest("/foo", handler.into_service())
+            .nest_service("/foo", handler.into_service())
             .layer(layer_fn(SetMatchedPathExtension));
 
         let client = TestClient::new(app);
@@ -198,12 +197,10 @@ mod tests {
     async fn nested_opaque_routers_append_to_matched_path() {
         let app = Router::new().nest(
             "/:a",
-            Router::new()
-                .route(
-                    "/:b",
-                    get(|path: MatchedPath| async move { path.as_str().to_owned() }),
-                )
-                .into_service(),
+            Router::new().route(
+                "/:b",
+                get(|path: MatchedPath| async move { path.as_str().to_owned() }),
+            ),
         );
 
         let client = TestClient::new(app);